# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Orders API
  version: 1.0.0
  description: API local de pedidos protegida por Bearer token.
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      operationId: health
      responses:
        '200': { description: Servicio disponible, content: { application/json: { schema: { $ref: '#/components/schemas/Health' } } } }
  /auth/login:
    post:
      operationId: login
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }
      responses:
        '200': { description: Credenciales válidas, content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /orders:
    get:
      operationId: listOrders
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 10 } }
      responses:
        '200': { description: Pedidos del usuario, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedOrders' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createOrder
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OrderRequest' } } } }
      responses:
        '201': { description: Pedido creado, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /orders/{id}:
    get:
      operationId: getOrder
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/OrderId' }]
      responses:
        '200': { description: Pedido del propietario, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /orders/{id}/status:
    put:
      operationId: setOrderStatus
      description: "Operación idempotente: repetir el mismo estado deja el mismo efecto final."
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/OrderId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StatusRequest' } } } }
      responses:
        '200': { description: Pedido actualizado, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer, bearerFormat: opaque } }
  parameters:
    OrderId: { name: id, in: path, required: true, schema: { type: string, pattern: '^ord-[a-zA-Z0-9-]+$' } }
  schemas:
    Health: { type: object, required: [status, service, version], properties: { status: { const: ok }, service: { type: string }, version: { type: string } }, additionalProperties: false }
    LoginRequest: { type: object, required: [username, password], properties: { username: { type: string }, password: { type: string, minLength: 1 } }, additionalProperties: false }
    LoginResponse: { type: object, required: [token, user], properties: { token: { type: string }, user: { $ref: '#/components/schemas/User' } }, additionalProperties: false }
    User: { type: object, required: [id, username, role], properties: { id: { type: string }, username: { type: string }, role: { const: customer } }, additionalProperties: false }
    OrderRequest: { type: object, required: [product, quantity], properties: { product: { type: string, minLength: 1, maxLength: 100 }, quantity: { type: integer, minimum: 1, maximum: 100 } }, additionalProperties: false }
    Order: { type: object, required: [id, ownerId, product, quantity, status, createdAt, updatedAt, owner], properties: { id: { type: string }, ownerId: { type: string }, product: { type: string }, quantity: { type: integer }, status: { enum: [pending, confirmed, cancelled] }, createdAt: { type: string, format: date-time }, updatedAt: { type: string, format: date-time }, owner: { enum: [me, other] } }, additionalProperties: false }
    PaginatedOrders: { type: object, required: [data, pagination], properties: { data: { type: array, items: { $ref: '#/components/schemas/Order' } }, pagination: { type: object, required: [page, pageSize, total, totalPages], properties: { page: { type: integer }, pageSize: { type: integer }, total: { type: integer }, totalPages: { type: integer } }, additionalProperties: false } }, additionalProperties: false }
    StatusRequest: { type: object, required: [status], properties: { status: { enum: [pending, confirmed, cancelled] } }, additionalProperties: false }
    Error: { type: object, required: [error], properties: { error: { type: object, required: [code, message, details], properties: { code: { type: string }, message: { type: string }, details: { type: array, items: { type: string } } }, additionalProperties: false } }, additionalProperties: false }
  responses:
    Unauthorized: { description: Token ausente o inválido, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Recurso ajeno, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Recurso inexistente, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    ValidationError: { description: Datos inválidos, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }